changed data types

+++ pkts.h
@@ -49,10 +49,10 @@ typedef struct {
   char mac_add[18];
   char essid[33];
 
-  int packet_count;
-  int n_enabled_count; // added on 14 feb, 2012
+  unsigned int packet_count;
+  unsigned int n_enabled_count; // added on 14 feb, 2012
 
-  u_int16_t bad_fcs_err_count;
+  unsigned int bad_fcs_err_count;
   u_int16_t short_preamble_err_count;
   u_int16_t radiotap_wep_err_count;
   u_int16_t retry_count;
@@ -61,8 +61,6 @@ typedef struct {
   u_int16_t retry_err_count;
   u_int16_t strictly_ordered_err_count;
 
-
-
   u_int16_t pwr_mgmt_count;
   u_int16_t wep_enc_count;
   u_int16_t more_frag_count;
@@ -70,8 +68,8 @@ typedef struct {
   float dbm_signal_sum;
   float dbm_noise_sum;
   
-  u_int8_t db_signal_sum;
-  u_int8_t db_noise_sum;
+  float db_signal_sum;
+  float db_noise_sum;
 
   u_int8_t channel;
   u_int8_t antenna; 
@@ -87,6 +85,7 @@ typedef struct {
   u_int8_t cap_privacy ;
   u_int8_t cap_ess_ibss ;
   u_int8_t channel_change;
+
 } address_table_entry_t;
 
 #define MAC_TABLE_ENTRIES 255
